Refactor: simplify github login
